% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fact.R, R/utils.R
\name{fact}
\alias{fact}
\alias{fact.default}
\alias{fact.logical}
\alias{fact.character}
\alias{fact.numeric}
\alias{fact.integer}
\alias{fact.complex}
\alias{fact.Date}
\alias{fact.POSIXct}
\alias{fact.POSIXlt}
\alias{fact.factor}
\alias{fact.fact}
\alias{values}
\alias{is_fact}
\alias{is.fact}
\title{Factor}
\usage{
fact(x)

\method{fact}{default}(x)

\method{fact}{logical}(x)

\method{fact}{character}(x)

\method{fact}{numeric}(x)

\method{fact}{integer}(x)

\method{fact}{complex}(x)

\method{fact}{Date}(x)

\method{fact}{POSIXct}(x)

\method{fact}{POSIXlt}(x)

\method{fact}{factor}(x)

\method{fact}{fact}(x)

values(x)

is_fact(x)

is.fact(x)
}
\arguments{
\item{x}{A vector of values}
}
\value{
A vector of equal length of `x` with class `fact` and `factor`.  If
  `x` was `ordered`, that class is added in between.
}
\description{
Quickly create a factor
}
\details{
[fact::fact()] can be about 5 times quicker than [base::factor()] or
  [base::as.factor()] as it doesn't bother sorting the levels for non-numeric
  data or have other checks or features.  It simply converts a vector to a
  factor with all unique values as levels with `NA`s included.

  [fact::fact.factor()] will perform several checks on a factor to include
  `NA` levels and to check if the levels should be reordered to conform with
  the other methods.  The [fact::fact.fact()] method simple returns `x`.

  By default, `NA` values are sorted as the first value.  This can be
  changed by setting `option(fact.na.position = "last")`.
}
\section{level orders}{


  The order of the levels may be adjusted to these rules depending on the
  class of `x`:
\describe{
  \item{`character`}{The order of appearance}
  \item{`numeric`/`integer`/`Date`/`POSIXt`}{By the numeric order}
  \item{`logical`}{As `FALSE`, `TRUE`}
  \item{`factor`}{Numeric if levels can be safely converted, otherwise as
  they are}
}
}

