% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fact.R
\name{fact_na}
\alias{fact_na}
\title{`fact` with `NA`}
\usage{
fact_na(x, remove = FALSE)
}
\arguments{
\item{x}{A `fact` or object cohered to `fact`}

\item{remove}{If `TRUE` removes `NA` value from the `fact` `levels` and
`uniques` attributes}
}
\value{
A `fact` vector
}
\description{
Included `NA` values into `fact()`
}
\details{
This re-formats the `x` value so that `NA`s are found immediately within the
object rather than accessed through its attributes.
}
\seealso{
Other factors: 
\code{\link{as_ordered}()},
\code{\link{drop_levels}()},
\code{\link{fact}()}
}
\concept{factors}
