% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facts-package.R
\docType{package}
\name{facts-package}
\alias{facts}
\alias{facts-package}
\title{facts: The 'fact' Class}
\description{
Implement and use the 'fact' class, an alternative to the 'factor' class, which tries to be faster and richer.
}
\keyword{internal}
