% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fact-utils.R
\name{is.fact}
\alias{is.fact}
\title{Is factor of class 'fact'?}
\usage{
is.fact(x)
}
\arguments{
\item{x}{Object to test}
}
\value{
Logical indicating if `x` is of class 'fact'
}
\description{
Is factor of class 'fact'?
}
