% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fact-methods.R
\name{levels.fact}
\alias{levels.fact}
\title{#' @export
c.fact <- function(...) {
  vec_c(...)
}}
\usage{
\method{levels}{fact}(x, ...)
}
\description{
#' @export
c.fact <- function(...) {
  vec_c(...)
}
}
