% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudo-id-methods.R
\name{print.pseudo_id}
\alias{print.pseudo_id}
\title{Print `pseudo_id`}
\usage{
\method{print}{pseudo_id}(x, ..., all = FALSE)
}
\arguments{
\item{x}{An object of class [pseudo_id]}

\item{...}{Not implemented}

\item{all}{if `TRUE` will print all values  This is not recommend for many
values as it will crowd the console output}
}
\value{
`x`, invisibly.  Called for its side effects.
}
\description{
Print `pseudo_id`
}
\seealso{
[pseudo_id()]
}
