% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudo-id.R
\name{pseudo_id}
\alias{pseudo_id}
\alias{pseudo_id.pseudo_id}
\alias{pseudo_id.default}
\alias{pseudo_id.factor}
\title{Create an ID for a vector}
\usage{
pseudo_id(x, ...)

\method{pseudo_id}{pseudo_id}(x, ...)

\method{pseudo_id}{default}(x, na_last = TRUE, ...)

\method{pseudo_id}{factor}(x, ...)
}
\arguments{
\item{x}{A vector of values}

\item{...}{Additional arguments passed to methods}

\item{na_last}{`Logical` if `FALSE` will not place `NA` at the end}
}
\value{
A `pseudo_id` object where the `integer` value of the vector
correspond to the position of the unique values in the attribute `"values"`.
}
\description{
Transforms a vector into an integer of IDs.
}
\examples{
set.seed(42)
(x <- sample(letters, 10, TRUE))
(pid <- pseudo_id(x))
attr(pid, "values")[pid]

}
