% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode.R
\name{recode}
\alias{recode}
\title{PLACEHOLDER}
\usage{
recode(x, ..., .list = list(...), .nomatch = c("na", "keep"))
}
\arguments{
\item{x, ..., .list, .nomatch}{PLACEHOLDER}
}
\description{
PLACEHOLDER
}
\examples{
# PLACEHOLDER

x <- fact(c(1, 1, 3, 0, 2))
recode(x, 1 ~ -1, 0 ~ NA)

}
