% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{as.ordered}
\alias{as.factor}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{generics}{\code{\link[generics:as.factor]{as.factor()}}, \code{\link[generics:as.ordered]{as.ordered()}}}
}}

