% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facts-package.R
\docType{package}
\name{facts-package}
\alias{facts}
\alias{facts-package}
\title{facts: The 'fact' Class}
\description{
Implement and use the facts class, a variant of the factors class,
}
\keyword{internal}
