% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facts-package.R
\docType{package}
\name{facts-package}
\alias{facts}
\alias{facts-package}
\title{facts: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\keyword{internal}
