% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facts-package.R
\docType{package}
\name{facts-package}
\alias{facts}
\alias{facts-package}
\title{facts: The 'fact' Class}
\description{
Implement and use the 'fact' class, an alternative to the 'factor' class, which tries to be faster and richer.
}
\author{
\strong{Maintainer}: Jordan Mark Barbone \email{jmbarbone@gmail.com} (\href{https://orcid.org/0000-0001-9788-3628}{ORCID}) [copyright holder]

}
\keyword{internal}
